added many Changes
create a trigger that activates on insert in book
log. created a trigger that activates on update on
book log. Added functionality to the delete user
button in the members page.

create a table copies which holds the details of
the available and total numbers of a certain books
The above mentioned triggers updates the available
column in copies.

Import the database before use;;;